% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.sapn_fit}
\alias{print.sapn_fit}
\title{Print a fitted SAPN model}
\usage{
\method{print}{sapn_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sapn_fit}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fitted SAPN model
}
